rename pub-loading-spinner to pub-spinner
