+++ components/productions/index/productions.vue
@@ -33,7 +33,7 @@
       </pub-panel-footer>
     </pub-panel>
     <div v-else-if="isLoading">
-      <pub-loading-spinner />
+      <pub-spinner />
     </div>
     <div v-else class="text-center">
       {{ $t('productions.overview.no_productions') }}
@@ -42,7 +42,7 @@
 </template>
 
 <script>
-  import PubLoadingSpinner from '@/publiq-ui/pub-loading-spinner';
+  import PubSpinner from '@/publiq-ui/pub-spinner';
   import PubPanel from '@/publiq-ui/pub-panel';
   import PubPanelFooter from '@/publiq-ui/pub-panel-footer';
   import PubPagination from '@/publiq-ui/pub-pagination';
@@ -52,7 +52,7 @@
 
   export default {
     components: {
-      PubLoadingSpinner,
+      PubSpinner,
       PubPanel,
       PubPanelFooter,
       PubPagination,
